+++ core/src/auto_enum/args.rs
@@ -91,7 +91,7 @@ pub(super) fn parse_args(args: TokenStream) -> Result<(Stack<Arg>, Marker, bool)
                     }
 
                     if never {
-                        Err(invalid_args!("multiple `never` option"))?;
+                        Err(arg_err!(i, "multiple `never` option"))?;
                     }
                     never = true;
                 }
@@ -100,9 +100,9 @@ pub(super) fn parse_args(args: TokenStream) -> Result<(Stack<Arg>, Marker, bool)
             TokenTree::Punct(p) => match p.as_char() {
                 ',' => {}
                 ':' => args.push(parse_path(smallvec![p.into()], &mut iter)?),
-                _ => Err(invalid_args!("{}`{}`", ERR, p))?,
+                _ => Err(arg_err!(p, "{}`{}`", ERR, p))?,
             },
-            _ => Err(invalid_args!("{}`{}`", ERR, tt))?,
+            _ => Err(arg_err!(tt, "{}`{}`", ERR, tt))?,
         }
     }
 
@@ -118,7 +118,7 @@ fn parse_path(mut path: Stack<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
     }
 
     syn::parse2(path.into_iter().collect())
-        .map_err(|e| invalid_args!(e))
+        .map_err(|e| arg_err!(e))
         .map(Arg::Path)
 }
 
@@ -130,9 +130,9 @@ fn path_or_ident(ident: Ident, tt: Option<TokenTree>, iter: &mut IntoIter) -> Re
         Some(TokenTree::Punct(p)) => match p.as_char() {
             ',' => Ok(ident.into()),
             ':' => parse_path(smallvec![ident.into(), p.into()], iter),
-            _ => Err(invalid_args!("{}`{}`", ERR, p)),
+            _ => Err(arg_err!(p, "{}`{}`", ERR, p)),
         },
-        Some(tt) => Err(invalid_args!("{}`{}`", ERR, tt)),
+        Some(tt) => Err(arg_err!(tt, "{}`{}`", ERR, tt)),
     }
 }
 
@@ -144,42 +144,41 @@ fn marker_opt(
 ) -> Result<()> {
     match iter.next() {
         Some(TokenTree::Group(ref g)) if g.delimiter() != Delimiter::Parenthesis => {
-            Err(invalid_args!("invalid delimiter"))?
+            Err(arg_err!(g, "invalid delimiter"))?
         }
         Some(TokenTree::Group(ref g)) => {
             let mut g = g.stream().into_iter();
             match g.next() {
-                Some(TokenTree::Ident(_)) if marker.is_some() => {
-                    Err(invalid_args!("multiple `marker` option"))?
+                Some(TokenTree::Ident(ref i)) if marker.is_some() => {
+                    Err(arg_err!(i, "multiple `marker` option"))?
                 }
                 Some(TokenTree::Ident(i)) => *marker = Some(i.to_string()),
-                Some(tt) => Err(invalid_args!("expected an identifier, found `{}`", tt))?,
-                None => Err(invalid_args!("empty `marker` option"))?,
+                Some(tt) => Err(arg_err!(tt, "expected an identifier, found `{}`", tt))?,
+                None => Err(arg_err!(ident, "empty `marker` option"))?,
             }
-
             match g.next() {
                 None => {}
                 Some(TokenTree::Punct(ref p)) if p.as_char() == ',' => {
-                    if g.next().is_some() {
-                        Err(invalid_args!("multiple identifier in `marker` option"))?;
+                    if let Some(tt) = g.next() {
+                        Err(arg_err!(tt, "multiple identifier in `marker` option"))?;
                     }
                 }
-                Some(_) => Err(invalid_args!("multiple identifier in `marker` option"))?,
+                Some(tt) => Err(arg_err!(tt, "multiple identifier in `marker` option"))?,
             }
         }
         Some(TokenTree::Punct(ref p)) if p.as_char() == '=' => {
             match iter.next() {
-                Some(TokenTree::Ident(_)) if marker.is_some() => {
-                    Err(invalid_args!("multiple `marker` option"))?
+                Some(TokenTree::Ident(ref i)) if marker.is_some() => {
+                    Err(arg_err!(i, "multiple `marker` option"))?
                 }
                 Some(TokenTree::Ident(i)) => *marker = Some(i.to_string()),
-                Some(tt) => Err(invalid_args!("expected an identifier, found `{}`", tt))?,
-                None => Err(invalid_args!("empty `marker` option"))?,
+                Some(tt) => Err(arg_err!(tt, "expected an identifier, found `{}`", tt))?,
+                None => Err(arg_err!(p, "empty `marker` option"))?,
             }
             match iter.next() {
                 None => {}
                 Some(TokenTree::Punct(ref p)) if p.as_char() == ',' => {}
-                Some(_) => Err(invalid_args!("multiple identifier in `marker` option"))?,
+                Some(tt) => Err(arg_err!(tt, "multiple identifier in `marker` option"))?,
             }
         }
         tt => args.push(path_or_ident(ident, tt, iter)?),